refactor: create bridge navigation hook
