+++ app/components/Views/WalletActions/WalletActions.tsx
@@ -18,6 +18,7 @@ import { toggleReceiveModal } from '../../../actions/modals';
 import { isSwapsAllowed } from '../../../components/UI/Swaps/utils';
 import isBridgeAllowed from '../../UI/Bridge/utils/isBridgeAllowed';
 import { useNavigation } from '@react-navigation/native';
+import useGoToBridge from '../../../components/UI/Bridge/utils/useGoToBridge';
 import Routes from '../../../constants/navigation/Routes';
 import { MetaMetricsEvents } from '../../../core/Analytics';
 import Analytics from '../../../core/Analytics/Analytics';
@@ -29,8 +30,6 @@ import WalletAction from '../../../components/UI/WalletAction';
 import { useStyles } from '../../../component-library/hooks';
 import generateTestId from '../../../../wdio/utils/generateTestId';
 import { AvatarSize } from '../../../component-library/components/Avatars/Avatar';
-import { BrowserTab } from '../../../components/UI/Tokens/types';
-import type { BrowserParams } from '../../../components/Views/Browser/Browser.types';
 
 // Internal dependencies
 import styleSheet from './WalletActions.styles';
@@ -48,8 +47,8 @@ const WalletActions = () => {
   const { styles } = useStyles(styleSheet, {});
   const sheetRef = useRef<SheetBottomRef>(null);
   const { navigate } = useNavigation();
+  const goToBridge = useGoToBridge('TabBar');
 
-  const browserTabs = useSelector((state: any) => state.browser.tabs);
   const chainId = useSelector(selectChainId);
   const ticker = useSelector(selectTicker);
   const swapsIsLive = useSelector(swapsLivenessSelector);
@@ -130,35 +129,6 @@ const WalletActions = () => {
     });
   };
 
-  const goToBridge = () => {
-    const bridgeUrl = `${AppConstants.PORTFOLIO_URL}/bridge`;
-    const existingBridgeTab = browserTabs.find((tab: BrowserTab) =>
-      tab.url.match(new RegExp(`${bridgeUrl}/(?![a-z])`)),
-    );
-
-    const params: BrowserParams & { existingTabId?: string } = {
-      timestamp: Date.now(),
-    };
-
-    if (existingBridgeTab) {
-      params.newTabUrl = undefined;
-      params.existingTabId = existingBridgeTab.id;
-    } else {
-      params.newTabUrl = `${bridgeUrl}/?metamaskEntry=mobile&srcChain=${chainId}`;
-    }
-
-    navigate(Routes.BROWSER.HOME, {
-      screen: Routes.BROWSER.VIEW,
-      params,
-    });
-    Analytics.trackEvent(MetaMetricsEvents.BRIDGE_LINK_CLICKED, {
-      bridgeUrl,
-      location: 'TabBar',
-      chain_id_source: chainId,
-      token_address_source: undefined,
-    });
-  };
-
   return (
     <SheetBottom ref={sheetRef}>
       <View style={styles.actionsContainer}>
@@ -206,7 +176,7 @@ const WalletActions = () => {
             actionDescription={strings('asset_overview.bridge_description')}
             iconName={IconName.Bridge}
             iconSize={AvatarSize.Md}
-            onPress={goToBridge}
+            onPress={() => goToBridge(chainId)}
             iconStyle={styles.icon}
             {...generateTestId(Platform, WALLET_BRIDGE)}
           />
